[NRRPLT-4219] Add VirtualCoach config.json and update setup.
This patch adds the config.json needed for the VirtualCoach build
and runtime and update the installation scripts to install the file.

It also includes a useful alias 'cle-virtual-coach' that can be used to
launch a python session where the virtual coach can be imported and used
or to run scripts that use the virtual coach (e.g. virtual-coach script.py).

Change-Id: I002410dfde7a910111d87cbbedb795898237df2c