[NRRPLT-5102] - Fix for the GET /populations fails for large brains
Le Fix:

This fix handles the issue that comes up whenever the frontend is doing 
a request for the .json containing the  populations in big brains. 
The problem was caused by the fact that the .json was bigger than what 
nginx would like to handle, so it was trying to use a cache in a file 
where the user running nginx (in this case us) does not have write 
permission. So the file was being only partially loaded and then an 
exception was being thrown. One solution would be to disable caching
completely by forbidding nginx to cache anything, another would be to 
change the ownership of the default cache folder. The cleanest solution
is to just change the caching location to /tmp , which is what 
this patch is doing. We already have nginx dumping other data in .tmp

Le test instructions:

- Install Alban's experiment 
https://bbpteam.epfl.ch/project/spaces/display/HSP10/Alban%27s+Visual+Segmentation+Experiment
- Click on the brainvisualizer icon, you should get an exception and 
on your developer tools an ERR::content_length_mismatch problem
- Apply le fix, restart nginx
- The brain visualizer should work fine

Change-Id: I0d26a94f5a7b19cc901bc79a13e277eb8c24d32a