[NRRPLT-6071] The update_nrp rebase script asks to stash based on the git-status
Before we were checking to see if there are unstaged or modified files
in the current branch preventing from switching to the dev or master
branch. The decision was based on a git diff-files command. Now the
decision is based on a git-status command where the untracked files
are ignored

Change-Id: Idbe4f57bb7305acca245502cd3bdfd91808a104f